requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==1.11.16
django-allauth==0.38.0
django-cors-headers==2.4.0
django-rest-auth==0.9.3
djangorestframework==3.9.0
djangorestframework-simplejwt==3.3
idna==2.7
oauthlib==2.1.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
PyJWT==1.7.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.24.1
whitenoise==3.3.1
